requirements.txt
pycord
python-a2s
tabulate
